double

+++ benchmarks/benchmark-fgemm.C
@@ -84,8 +84,8 @@ int main(int argc, char** argv) {
     //  typedef Givaro::Modular<double> Field;
     //  typedef Givaro::Modular<float> Field;
     //  typedef Givaro::ModularBalanced<float> Field;
-    //typedef Givaro::ModularBalanced<double> Field;
-    typedef Givaro::ModularBalanced<int64_t> Field;
+    typedef Givaro::ModularBalanced<double> Field;
+    //  typedef Givaro::ModularBalanced<int64_t> Field;
     //  typedef Givaro::Modular<Givaro::Integer> Field;
     typedef Field::Element Element;
